+++ changelog.md
@@ -3,10 +3,11 @@
 * **Personal Loot**
 * RCLootCouncil is now fully useable with Personal Loot (PL).
 * All features (except obviously automatic distribution) is available when using PL.
-* PL mode can be activated the same way ML mode is activated (popup in raid/group and/or through options).
+* PL mode is activated the same way ML used to be (popup in raid/group and/or through options).
 * When in PL mode, any tradeable items looted by any raider is automatically added to the session frame.
 * When items are awarded, players will need to trade the item to the winner, see below.
 * PL sessions still respects all loot/item settings such as autopass, filters etc.
+* *TODO: Missing a list of candidates that has looted a particular boss.*
 
 
 * **TradeUI**
@@ -24,6 +25,7 @@
 * **Item owners**
 * Session- and Voting frame now shows the owner of a particular item.
 * This info is also available for use in announcements.
+
 
 * **Discord export**
 * The Loot History can now be exported in a Discord friendly format.